+++ game/views/common.py
@@ -41,6 +41,9 @@ class GameView(LoginRequiredMixin, ListView, gmixins.GameContextMixin):
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         context["tale"] = gmodels.Tale.objects.filter(game=self.game.id).last()
+        context["instruction"] = gmodels.Instruction.objects.filter(
+            game=self.game.id
+        ).last()
         context["character_list"] = cmodels.Character.objects.filter(
             player__game=self.game.id
         ).order_by("name")